Move HowTo component and its related files to the features/HowTo directory
