+++ frontend/src/features/HowTo/components/Smartphone/Smartphone.tsx
