+++ frontend/src/features/HowTo/components/Smartphone/index.ts
